ui: move settings components into correct dir
